Fix types used in SMC code

+++ src/arch/aarch64/smc.h
@@ -7,12 +7,13 @@
 
 #pragma once
 
-#include <stdint.h>
 #include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
 #include <sel4cp.h>
 
 /* SMC vCPU fault handler */
-bool handle_smc(uint64_t vcpu_id, uint32_t hsr);
+bool handle_smc(size_t vcpu_id, uint32_t hsr);
 
 /* Helper functions */
 void smc_set_return_value(seL4_UserContext *u, uint64_t val);
